3061 - Small refactor
