+++ src/meta/tablePaginated/index.ts
@@ -1,10 +1,4 @@
 export type { TablePaginatedCompareFn } from './sortFn'
-export type {
-  InvitationFilters,
-  TablePaginatedCount,
-  TablePaginatedFilterValues,
-  TablePaginatedOrderBy,
-  UserFilters,
-} from './tablePaginated'
+export type { TablePaginatedCount, TablePaginatedFilterValues, TablePaginatedOrderBy } from './tablePaginated'
 export { TablePaginatedFilterType, TablePaginatedOrderByDirection } from './tablePaginated'
 export { TablePaginateds } from './tablePaginateds'
